code refactor

+++ src/main/java/com/speedway/api/APIRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface APIRepository extends JpaRepository<Student, Long> {
+public interface APIRepository extends JpaRepository<Student, Integer> {
 }
